refactor: extract input selection helpers and tests
